# backend/openapi/openapi.yaml
openapi: 3.0.4
info:
  title: Task Management API
  description: API for Task Management software backend
  version: 0.0.1

servers:
  - url: http://api.jamesbarnfather.co.uk
    description: Main (Production) server

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  responses:
    UnauthorizedError:
      description: Access token is missing or invalid

  schemas:
    RegisterSchema:
      type: object
      required:
        - email
        - password
        - username
      properties:
        email:
          type: string
          format: email
        password:
          type: string
        username:
          type: string

    LoginSchema:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
        password:
          type: string

    AuthenticationResponse:
      type: object
      properties:
        access:
          type: string
        refresh:
          type: string

    RefreshSchema:
      type: object
      required:
        - refresh
      properties:
        refresh:
          type: string

    UserResponse:
      type: object
      properties:
        email:
          type: string
        username:
          type: string
        role:
          type: string
        createdAt:
          type: string

    ProjectSchema:
      type: object
      required:
        - name
        - description
      properties:
        name:
          type: string
        description:
          type: string
    ProjectResponse:
      type: object
      required:
        - id
      properties:
        id:
          type: number

    TaskSchema:
      type: object
      required:
        - projectId
        - title
      properties:
        projectId:
          type: string
        title:
          type: string
        description:
          type: string
        status:
          type: string
          enum: [TODO, IN_PROGRESS, DONE]
    TaskResponse:
      type: object
      required:
        - id
      properties:
        id:
          type: number

paths:
  /:
    get:
      summary: Root
      responses:
        "200":
          description: Succesfully registered
          content:
            text/plain:
              schema:
                type: string
                example: pong
  /auth/register:
    post:
      summary: Register an account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterSchema"
      responses:
        "200":
          description: Succesfully registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthenticationResponse"
        "409":
          description: Account already exists
  /auth/login:
    post:
      summary: Login to an account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginSchema"
      responses:
        "200":
          description: Succesfully logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthenticationResponse"
        "401":
          description: Invalid Credentials exists
  /auth/logout:
    post:
      summary: Logout
      responses:
        "204":
          description: Successfully logout
  /auth/refresh:
    post:
      summary: Refresh authentication tokens
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RefreshSchema"
      responses:
        "200":
          description: Succesfully logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthenticationResponse"

  /user:
    get:
      summary: Get user information
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Returns user data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserResponse"
        "401":
          $ref: "#/components/responses/UnauthorizedError"

  /projects:
    get:
      summary: Get all projects linked with user
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Returns user data (TODO change to proper response)
          content:
            text/plain:
              schema:
                type: string
                example: pong
        "401":
          $ref: "#/components/responses/UnauthorizedError"
    post:
      summary: Create new project for user
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProjectSchema"
      responses:
        "200":
          description: Succesfully created project
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectResponse"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
  /projects/{projectId}:
    get:
      summary: Get project by ID
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: projectId
          schema:
            type: integer
          required: true
          description: Numeric ID of the project you want to get
      responses:
        "200":
          description: Returns user data (TODO change to proper response)
          content:
            text/plain:
              schema:
                type: string
                example: pong
        "401":
          $ref: "#/components/responses/UnauthorizedError"
    post:
      summary: Update project by ID
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: projectId
          schema:
            type: integer
          required: true
          description: Numeric ID of the project you want to get
      responses:
        "204":
          description: OK
        "401":
          $ref: "#/components/responses/UnauthorizedError"
    delete:
      summary: Delete project by ID
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: projectId
          schema:
            type: integer
          required: true
          description: Numeric ID of the project you want to get
      responses:
        "204":
          description: Project was deleted
        "401":
          $ref: "#/components/responses/UnauthorizedError"

  /tasks:
    post:
      description: ""
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TaskSchema"
      responses:
        "200":
          description: OK
        "401":
          $ref: "#/components/responses/UnauthorizedError"
  /tasks/{projectId}:
    get:
      description: Get all tasks tied to project
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: projectId
          schema:
            type: integer
          required: true
          description: Numeric ID of the project that you want to retrieve tasks from
      responses:
        "200":
          description: Returns task ID
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TaskResponse"
        "401":
          $ref: "#/components/responses/UnauthorizedError"